+++ app/widgets/Rooms/Rooms.php
@@ -42,6 +42,7 @@ class Rooms extends Base
         $this->registerEvent('presence_muc_errorconflict', 'onConflict');
         $this->registerEvent('presence_muc_errorregistrationrequired', 'onRegistrationRequired');
         $this->registerEvent('presence_muc_errorremoteservernotfound', 'onRemoteServerNotFound');
+        $this->registerEvent('presence_muc_errornotauthorized', 'onNotAuthorized');
         $this->registerEvent('composing', 'onComposing', 'chat');
         $this->registerEvent('paused', 'onPaused', 'chat');
 
@@ -101,6 +102,12 @@ class Rooms extends Base
     public function onRemoteServerNotFound($packet)
     {
         Notification::toast($this->__('chatrooms.remoteservernotfound'));
+        $this->ajaxExit($packet->content);
+    }
+
+    public function onNotAuthorized($packet)
+    {
+        Notification::toast($this->__('chatrooms.notauthorized'));
         $this->ajaxExit($packet->content);
     }