+++ lib/moxl/src/Moxl/Xec/Action/Presence/Muc.php
@@ -101,6 +101,12 @@ class Muc extends Action
         $this->deliver();
     }
 
+    public function errorNotAuthorized($stanza, $parent = false)
+    {
+        $this->pack($this->_to);
+        $this->deliver();
+    }
+
     public function errorConflict($stanza, $message)
     {
         if (substr_count($this->_nickname, '_') > 5) {